feature: created banner.jsx and banner.scss
