+++ kasa-immo/src/components/Banner/Banner.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Banner() {
+    return (
+        <div className="banner">
+            Banner
+        </div>
+    )
+}
+
+export default Banner;
