+++ kasa-immo/src/layout/Header/Header.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import LogoIcon from "../../components/LogoIcon";
+import LogoIcon from "../../components/LogoIcon/LogoIcon";
 import '../../styles/layout/_header.scss'
 
 function Header() {
