+++ kasa-immo/src/styles/components/banner.scss
@@ -1,0 +1,1 @@
+.banner {}
